+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All [Semantic Versions](https://semver.org/spec/v2.0.0.html) of this project and
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), with
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Tab and Shift+Tab keyboard shortcuts for navigating between TUI tabs (e498)
+
 ## [0.6.0-rc.1] - 2025-07-11
 
 ### Added